move streaming filter to separate package
